+++ ArcadiaParties-SPA/src/app/department/models/department.ts
